requirements.txt
Flask
Werkzeug
pdfplumber
pymongo
flask-cors
numpy
llama-index
llama-index-embeddings-voyageai
llama-index-storage-docstore-mongodb
llama-index-storage-index-store-mongodb
llama-index-llms-openai
llama-index-llms-ollama
torch
faiss-cpu
tenacity
pyjwt
streamlit
python-dotenv
